backup/bak_20260308_1345/requirements.txt
PySide6>=6.6.0
openpyxl>=3.1.0
python-docx>=1.0.0
# sqlite3 ist in Python standardmäßig enthalten
